+++ HISTORY.md
@@ -3,6 +3,7 @@
 ## v0.1.0: Unreleased
 
 - Renamed to `layouts`
+- Added inherited layout support
 
 ## v0.0.1: 2015-07-03